feat: Update calendar container styling
The code changes in `home.css` and `home.scss` update the styling of the calendar container. The `.calendar-container` class now has a `display` property set to `flex`, `justify-content` set to `center`, and a margin of `50px` on the top and bottom. This change improves the layout and alignment of the calendar on the page.